+++ src/main.rs
@@ -165,10 +165,12 @@ async fn run(
         .context("Failed to initialize last timestamp data file")?;
     read_last_timestamp(&settings.last_timestamp_data).await?;
 
+    let giganto_name = settings.resolve_giganto_name()?;
+    let giganto_publish_srv_addr = settings.giganto_publish_srv_addr()?;
     let subscribe_client = subscribe::Client::new(
         settings.giganto_ingest_srv_addr,
-        settings.giganto_publish_srv_addr,
-        settings.giganto_name,
+        giganto_publish_srv_addr,
+        giganto_name,
         settings.last_timestamp_data,
         certs,
         request_recv,